% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS.R
\name{CI}
\alias{CI}
\title{Confidence Intervals for OLS model}
\usage{
CI(model, variables, level = 0.95)
}
\arguments{
\item{model}{a fitted model by OLS function.}

\item{variables}{a vector specifying which parameters are to be given intervals.
The default indicator for intercept is Intercept.}

\item{level}{The confidence level required. The default level is 0.95.}
}
\value{
A matrix with columns giving lower and upper confidence limits for each
selected parameter.
}
\description{
Computes confidence intervals for parameters in model fitted by OLS
function'
}
\examples{
## Continue with OLS example
outcome = "mpg"
predictors = c("cyl", "hp", "wt")
m1 = OLS(mtcars, outcome, predictors, through_origin=FALSE)
CI(m1, c("intercept", "cyl", "hp", "wt"), level=0.95)
CI(m1,"cyl",level=0.9)

}
