% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols-package.R
\docType{package}
\name{ols-package}
\alias{ols}
\alias{ols-package}
\title{ols: REST API Client for the Ontology Lookup Service}
\description{
Programmatic access to the Ontology Lookup Service (OLS). This package provides easy access to biomedical ontologies as made available by the European Bioinformatics Institute 'OLS' through its REST API \url{https://www.ebi.ac.uk/ols4/api/ontologies}.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
