% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS.R
\name{OLS}
\alias{OLS}
\title{Ordinary Least Square Method}
\usage{
OLS(data, outcome, predictors, through_origin = FALSE)
}
\arguments{
\item{data}{an object of data frame or list containing all variables and
observations in the model}

\item{outcome}{an vector specifying an outcome in the model. It should be a
string or an integer to indicate the location of outcome in data.}

\item{predictors}{an vector specifying predictors in the model. It should be
string or integers to indicate the location of outcome in data.}

\item{through_origin}{logical. If TRUE the model is forced to go through
origin. The default is FALSE.}
}
\value{
A list of results of fitted model. design_matrix: The design matrix used.
response: The response used. coefficients: a named vector of estimated coefficients.
residuals: The response minus fitted values. df: The degree of freedom of residuals.
through_origin: A logical indicator of whether the model pass the origin.
variance: A vector containing basic analysis of variance of fitted model,
including SSE, SSR, SST and R squared statistics. summary: Summary of the
model results, containing estimates, results of hypothesis test (p value) and
fitting results (R-square statistics).
}
\description{
OLS is used to fit linear models using ordinary least square method.
It can be used to carry out estimates and hypothesis tests for each parameter
in linear regression. A simple analysis of variance is also provided but can
be examined further by AOV function in this package.'
}
\examples{
## Build-in data set mtcars in R
outcome = "mpg"
predictors = c("cyl", "hp", "wt")
m1 = OLS(mtcars, outcome, predictors, through_origin=FALSE)
m2 = OLS(mtcars, outcome, predictors, through_origin=TRUE) # omitting intercept


}
